Breaker: it's not an iOS app only; can be found on Android.
gPodder: it's not a Windows app only; can be found on other OS.

Himalaya: it's not an iOS app only; can be found on Android.

^AppleCoreMedia/1.(.*)HomePod: the OS is HomePod OS, not iOS.

Player FM: it's not an Android app only; can be found on iOS.

Spreaker: it's not an Android app only; can be found on iOS.

TuneIn: it's not an iOS app only; can be found on Android.

yapa: it's not an Android app only; can be found on iOS.