Simplifies regex
This puts the onus on implementers to parse regexes to match the quirks in their own system, prioritising the human-editability of this document.

eg: slightly quirky interpretations of regex put greater importance on the forward-slash (“/“) character than is always useful, but those using those systems know this is the case, so can code around them, whereas those editing this document shouldn’t have to think of esoteric regex implementation details.